Add bare bones `wigpay` Spring Boot application.
Started with a web project created by the Spring Initializr.  Added a
simple handler for `/` to return "Hello {name}" as a test file.  Added
classes from `igpay` and `TranslationController` class with a handler
for `/translation` that takes an `english` parameter and returns a
simple JSON response created from the `Translation` record class.